+++ examples/coffeescript/gulpfile.js
@@ -34,7 +34,8 @@ const coffeeScript = {
     // TODO May 6, 2018
     // If transpile option is given, gulp-coffee fails when sourcemaps are enabled.
     // No solution found, so this option is blocked until the solution is found.
-    // coffee: {transpile: {"presets": ["env"]}}
+    // gulp-coffee: v3.0.2, issue #91
+    // coffee: {transpile: {presets: ['env']}}
   }
 };